remove legacy Python OrderGroup and MultiFeedClock helpers
Pre-release we don't keep two parallel APIs. The Python-only
flox_py.execution.OrderGroup and flox_py.feed_clock.MultiFeedClock
helpers existed only as a stepping stone before the engine-backed
versions landed. The engine-backed primitives (flox_py.OrderGroup
and flox_py.MultiFeedClock) reach every binding through the C ABI,
expose the full policy surface, and are strictly better.

- Deletes python/flox_py/execution.py and the legacy OrderGroup /
  GroupPolicy / GroupStatus exports.
- Deletes python/flox_py/feed_clock.py and the legacy MultiFeedClock
  / Policy / ClockState Python class.
- Deletes the legacy test files (test_order_group.py,
  test_feed_clock.py).
- Renames test_engine_order_group.py to test_order_group.py and
  test_engine_feed_clock.py to test_feed_clock.py; drops the
  "legacy" docstring comparisons.
- Rewrites docs/how-to/multi-leg-order-group.md and
  docs/how-to/multi-feed-clock.md to describe a single API rather
  than "two surfaces".

W14-T013
