+++ docs/how-to/multi-feed-clock.md
@@ -2,64 +2,45 @@
 
 Cross-symbol decisions read from N feeds. Without a clock, a strategy that recomputes a ratio on every BTC tick uses a stale ETH price every time the ETH feed lags. The `MultiFeedClock` makes the staleness budget explicit and exposes per-feed lag so the strategy can score its own decisions.
 
-```python
-from flox_py.feed_clock import MultiFeedClock, WaitForAll
-
-class PairTrade(Strategy):
-    def setup(self):
-        self.clock = MultiFeedClock(
-            symbols=[btc_id, eth_id],
-            policy=WaitForAll,
-            timeout_ms=200,
-        )
-
-    def on_trade(self, ctx, trade):
-        state = self.clock.tick(trade.exchange_ts_ns, ctx.symbol_id)
-        if state.fired:
-            self.evaluate(state)
-```
-
-## Policies
-
-- `WaitForAll`: fire when every listed feed has emitted at least one event since the last fire. If the timeout elapses first, fire anyway with whatever the latest known prices are; `state.staleness_ns` lets the strategy down-weight the decision.
-- `FireOnAny`: fire on every tick from any listed feed (the original default). Explicit so the staleness budget is visible to readers.
-- `LeaderFollower`: fire on the leader's tick only if every follower's lag is within `staleness_budget_ms`. Used when one feed is the source of truth and the others are confirmation.
-
-## ClockState
-
-`tick()` returns a `ClockState` with:
-
-- `fired`: whether the strategy should evaluate now.
-- `last_ts_ns[symbol]`: most recent observed timestamp per feed.
-- `staleness_ns[symbol]`: `tick_ts - last_ts` per feed at the moment of fire.
-- `triggered_by`: which feed's tick caused the fire (or `None` for out-of-band).
-
-## Out-of-band symbols
-
-Calling `tick()` with a symbol that was not in the original `symbols` list updates the per-symbol last-seen timestamp but never causes a fire on its own. This keeps the clock honest if a strategy is also subscribed to feeds beyond the cross-symbol decision.
-
-## Tests
-
-`python/tests/test_feed_clock.py` covers all three policies, the timeout fallback path, the staleness map, the leader / follower freshness check, and the out-of-band symbol case.
-
-## Engine primitive
-
-The doc above describes `flox_py.feed_clock.MultiFeedClock` — a pure-Python helper that ships with flox_py. There is also a C++ engine primitive (`include/flox/feed/multi_feed_clock.h`) reachable from every binding through a shared C ABI. The two have the same shape; pick the engine version when you need polyglot parity or a zero-overhead hot path.
-
-The engine primitive exposes named string constants for the policy in JS bindings and an enum (`flox_py.FeedClockPolicy`) in Python so users never hand-write integer codes.
+The clock lives in `include/flox/feed/multi_feed_clock.h` as a C++ primitive. Every binding (pybind11 / NAPI / QuickJS / Codon) wraps the same C ABI so a strategy in any language gets the same semantics and the same hot-path cost. Policies are exposed as named string constants in JS bindings and as an enum (`flox_py.FeedClockPolicy`) in Python — users never hand-write integer codes.
 
 ```python
 # pybind11
 import flox_py
+
 clock = flox_py.MultiFeedClock(
     symbols=[btc, eth],
     policy=flox_py.FeedClockPolicy.WAIT_FOR_ALL,
     timeout_ms=200,
 )
-state = clock.tick(ts_ns, btc)
-if state["fired"]:
-    print(state["staleness_ns"])  # {btc: 0, eth: 1234567}
+
+class PairTrade(Strategy):
+    def on_trade(self, ctx, trade):
+        state = clock.tick(trade.exchange_ts_ns, ctx.symbol_id)
+        if state["fired"]:
+            self.evaluate(state)
 ```
+
+## Policies
+
+- `WaitForAll`: fire when every listed feed has emitted at least one event since the last fire. If the timeout elapses first, fire anyway with whatever the latest known prices are; the per-feed `staleness_ns` map lets the strategy down-weight the decision.
+- `FireOnAny`: fire on every tick from any listed feed. Explicit so the staleness budget is visible to readers.
+- `LeaderFollower`: fire on the leader's tick only if every follower's lag is within `staleness_budget_ms`. Used when one feed is the source of truth and the others are confirmation.
+
+## Tick result
+
+`tick(ts_ns, symbol)` returns a snapshot with:
+
+- `fired`: whether the strategy should evaluate now.
+- `last_ts_ns[symbol]`: most recent observed timestamp per feed.
+- `staleness_ns[symbol]`: `tick_ts - last_ts` per feed at the moment of fire.
+- `triggered_by`: which feed's tick caused the fire (or `0` for out-of-band).
+
+## Out-of-band symbols
+
+Calling `tick()` with a symbol that was not in the original `symbols` list updates the per-symbol last-seen timestamp but never causes a fire on its own. This keeps the clock honest if a strategy is also subscribed to feeds beyond the cross-symbol decision.
+
+## Cross-binding
 
 ```javascript
 // node
@@ -88,3 +69,7 @@ from flox.feed_clock import MultiFeedClock, WAIT_FOR_ALL
 clock = MultiFeedClock(symbols=[btc, eth], policy=WAIT_FOR_ALL, timeout_ms=200)
 state = clock.tick(ts_ns, btc)
 ```
+
+## Tests
+
+`python/tests/test_feed_clock.py` covers all three policies, the timeout fallback path, the staleness map, the leader / follower freshness check.
