--- python/tests/test_engine_order_group.py
@@ -1,71 +1,0 @@
-"""Tests for the C++-engine-backed OrderGroup state machine (W15-T004).
-
-The legacy Python helper lives in `flox_py.execution.OrderGroup`. This
-suite exercises the new `flox_py.OrderGroup` (bound from
-`include/flox/execution/order_group.h`), which the NAPI / QuickJS /
-Codon bindings share through the C ABI.
-"""
-from __future__ import annotations
-
-import pytest
-
-import flox_py
-
-
-def test_best_effort_recommends_nothing() -> None:
-    g = flox_py.OrderGroup(parent_signal_id=1, policy=flox_py.OrderGroupPolicy.BEST_EFFORT)
-    a = g.add_market_leg(symbol=1, side=0, qty=0.1)
-    b = g.add_market_leg(symbol=2, side=1, qty=2.0)
-    assert g.leg_count() == 2
-    assert g.state() == flox_py.OrderGroupState.PENDING
-    g.record_submit(a, 100)
-    g.record_submit(b, 101)
-    assert g.state() == flox_py.OrderGroupState.SUBMITTED
-    g.record_fill(a, 0.1)
-    g.record_failure(b)
-    assert g.state() == flox_py.OrderGroupState.PARTIALLY_FILLED
-    assert g.recommended_actions() == []
-
-
-def test_one_sided_cancels_remaining_on_first_fill() -> None:
-    g = flox_py.OrderGroup(policy=flox_py.OrderGroupPolicy.ONE_SIDED)
-    g.add_limit_leg(symbol=1, side=0, price=50000.0, qty=0.1)
-    g.add_limit_leg(symbol=2, side=1, price=3000.0, qty=1.5)
-    g.record_submit(0, 200)
-    g.record_submit(1, 201)
-    g.record_fill(0, 0.1)
-
-    actions = g.recommended_actions()
-    assert len(actions) == 1
-    assert actions[0]["kind"] == "cancel"
-    assert actions[0]["leg_index"] == 1
-    assert actions[0]["order_id"] == 201
-
-
-def test_all_or_nothing_reverts_filled_legs_on_failure() -> None:
-    g = flox_py.OrderGroup(policy=flox_py.OrderGroupPolicy.ALL_OR_NOTHING)
-    g.add_market_leg(1, 0, 0.1)
-    g.add_market_leg(2, 1, 2.0)
-    g.record_submit(0, 300)
-    g.record_submit(1, 301)
-    g.record_fill(0, 0.1)
-    g.record_failure(1)
-
-    assert g.state() == flox_py.OrderGroupState.REVERTING
-    actions = g.recommended_actions()
-    assert len(actions) == 1
-    assert actions[0]["kind"] == "revert"
-    assert actions[0]["symbol"] == 1
-    # Opposite side of the original BUY (0) → SELL (1)
-    assert actions[0]["side"] == 1
-    assert actions[0]["qty"] == pytest.approx(0.1)
-
-
-def test_partial_fill_marks_leg_partially_filled() -> None:
-    g = flox_py.OrderGroup(policy=flox_py.OrderGroupPolicy.BEST_EFFORT)
-    g.add_market_leg(1, 0, 0.5)
-    g.record_submit(0, 400)
-    g.record_fill(0, 0.2)  # half of target
-    assert g.leg_state(0) == flox_py.LegState.PARTIALLY_FILLED
-    g.record_fill(0, 0.5)
-    assert g.leg_state(0) == flox_py.LegState.FILLED
